PSG-1710: move reheader_fasta.py script to scripts/common/
